+++ lib/td_audit_web/views/email_view.ex
@@ -35,7 +35,7 @@ defmodule TdAuditWeb.EmailView do
       name: EventView.resource_name(event),
       values: values,
       domains: domain_path(event),
-      date: payload["date"],
+      date: TdAudit.Helpers.shift_zone(payload["date"]),
       uri: uri(event)
     )
   end